+++ src/main/java/mod/emt/thaumictweaker/mixins/focus/FocusEffectHealMixin.java
@@ -1,12 +1,17 @@
 package mod.emt.thaumictweaker.mixins.focus;
 
 import mod.emt.thaumictweaker.config.ConfigEnhancementsTT;
+import mod.emt.thaumictweaker.config.ConfigOverhaulsTT;
 import net.minecraft.entity.Entity;
+import net.minecraft.entity.EntityLivingBase;
+import net.minecraft.init.MobEffects;
+import net.minecraft.potion.PotionEffect;
 import net.minecraft.util.SoundCategory;
 import net.minecraft.util.math.RayTraceResult;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.Redirect;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 import thaumcraft.api.casters.FocusEffect;
 import thaumcraft.api.casters.Trajectory;
@@ -22,7 +27,26 @@ public abstract class FocusEffectHealMixin extends FocusEffect {
         try {
             if(ConfigEnhancementsTT.enableFocusEffects)
                 caster.world.playSound(null, caster.getPosition().up(), SoundsTC.wand, SoundCategory.PLAYERS, 0.825F, 3.0F + (float) (caster.world.rand.nextGaussian() * 0.05F));
-        } catch (Exception ignored) {
+        } catch(Exception ignored) {
+        }
+    }
+
+    @Redirect(
+            method = "execute",
+            at = @At(
+                    value = "INVOKE",
+                    target = "Lnet/minecraft/entity/EntityLivingBase;heal(F)V"
+            ),
+            remap = false
+    )
+    public void replaceHealWithRegen(EntityLivingBase entity, float amount) {
+        if(ConfigOverhaulsTT.nerfedHealFocus) {
+            int duration = (1 + this.getSettingValue("power")) * 20;
+            if(MobEffects.REGENERATION != null) {
+                entity.addPotionEffect(new PotionEffect(MobEffects.REGENERATION, duration, 2, true, true));
+            }
+        } else {
+            entity.heal(amount);
         }
     }
 
@@ -31,7 +55,7 @@ public abstract class FocusEffectHealMixin extends FocusEffect {
         try {
             if(ConfigEnhancementsTT.enableFocusEffects)
                 this.getPackage().world.playSound(null, target.hitVec.x, target.hitVec.y, target.hitVec.z, SoundsTC.wand, SoundCategory.PLAYERS, 0.525F, 0.7F + (float) (this.getPackage().getCaster().world.rand.nextGaussian() * 0.05F));
-        } catch (Exception ignored) {
+        } catch(Exception ignored) {
         }
     }