+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 - Fully merged Thaumic Inventory Scanning into the mod with major improvements
 - Restored Flux Rift drain fix, it has been removed from Thaumcraft Fix since v1.1.10
+- Added Nerfed Heal Focus Effect overhaul to prevent the Heal Focus Effect from being too overpowered
 ### Changed
 - Modified `TileVoidSiphonMixin` to support upcoming ThaumicAPI handling